+++ app/src/main/java/io/github/sds100/keymapper/actions/GetActionErrorUseCase.kt
@@ -7,6 +7,7 @@ import io.github.sds100.keymapper.system.camera.CameraAdapter
 import io.github.sds100.keymapper.system.inputmethod.InputMethodAdapter
 import io.github.sds100.keymapper.system.permissions.PermissionAdapter
 import io.github.sds100.keymapper.system.permissions.SystemFeatureAdapter
+import io.github.sds100.keymapper.system.ringtones.RingtoneAdapter
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.channelFlow
 import kotlinx.coroutines.flow.collectLatest
@@ -22,6 +23,7 @@ class GetActionErrorUseCaseImpl(
     private val cameraAdapter: CameraAdapter,
     private val soundsManager: SoundsManager,
     private val shizukuAdapter: ShizukuAdapter,
+    private val ringtoneAdapter: RingtoneAdapter,
 ) : GetActionErrorUseCase {
 
     private val invalidateActionErrors = merge(
@@ -51,6 +53,7 @@ class GetActionErrorUseCaseImpl(
         cameraAdapter,
         soundsManager,
         shizukuAdapter,
+        ringtoneAdapter,
     )
 }