+++ app/src/main/java/io/github/sds100/keymapper/util/FilterUtils.kt
@@ -11,7 +11,7 @@ import java.util.Locale
  * Created by sds100 on 22/03/2021.
  */
 
-suspend fun <T : ISearchable> List<T>.filterByQuery(query: String?): Flow<State<List<T>>> = flow {
+fun <T : ISearchable> List<T>.filterByQuery(query: String?): Flow<State<List<T>>> = flow {
     if (query.isNullOrBlank()) {
         emit(State.Data(this@filterByQuery))
     } else {